requirements.txt
python-magic; platform_system != "Windows"
python-magic-bin==0.4.14; platform_system == "Windows"
